split off state.py
